# perf(autolinking): get platform's specific properties