@szymonrybczak 2024-05-07
CI Blocked by https://github.com/react-native-community/cli/pull/2376